pyproject.toml
[tool.poetry]
name = "bd-interface"
version = "0.1.0"
description = "rest api to manage mongodb"
authors = ["jordaoa"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask-pydantic-spec = "0.4.5"
Flask = "2.3"
mongoengine = "0.27.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
